moved alignment/line-spacing defatult data into buttonlistcreatorconsts
