--- packages/toolbars-new/web/src/modals/alignment/consts.ts
@@ -1,28 +1,0 @@
-import { AlignTextCenterIcon, AlignJustifyIcon, AlignLeftIcon, AlignRightIcon } from '../../icons';
-
-export default [
-  {
-    text: 'AlignObject_Left_Tooltip',
-    tooltip: 'AlignObject_Left_Tooltip',
-    commandKey: 'left',
-    icon: AlignLeftIcon,
-  },
-  {
-    text: 'AlignObject_Center_Tooltip',
-    tooltip: 'AlignObject_Center_Tooltip',
-    commandKey: 'center',
-    icon: AlignTextCenterIcon,
-  },
-  {
-    text: 'AlignObject_Right_Tooltip',
-    tooltip: 'AlignObject_Right_Tooltip',
-    commandKey: 'right',
-    icon: AlignRightIcon,
-  },
-  {
-    text: 'AlignTextJustifyButton_Tooltip',
-    tooltip: 'AlignTextJustifyButton_Tooltip',
-    commandKey: 'justify',
-    icon: AlignJustifyIcon,
-  },
-];
