--- packages/toolbars-new/web/src/modals/line-spacing/consts.ts
@@ -1,6 +1,0 @@
-export default [
-  { text: '1', commandKey: '1' },
-  { text: '2', commandKey: '2' },
-  { text: '2.5', commandKey: '2.5' },
-  { text: '3', commandKey: '3' },
-];
